requirements.txt
# AI 분석 파이프라인 의존성

# Whisper 음성 인식
openai-whisper>=20240930

# YouTube 다운로드
yt-dlp>=2024.12.23

# Google Generative AI
google-generativeai>=0.8.3

# Google API Client (YouTube Data API v3)
google-api-python-client>=2.134.0
google-auth>=2.30.0
google-auth-httplib2>=0.2.0

# 데이터 검증
pydantic>=2.10.0

# 이미지 처리 및 OCR
Pillow>=10.4.0
easyocr>=1.7.0
opencv-python>=4.8.0
numpy>=1.24.0

# 객체 인식 (향후 YOLO용)
ultralytics>=8.3.0

# GPU 최적화 및 딥러닝
torch>=2.0.0
torchvision>=0.15.0
nvidia-ml-py3>=7.352.0

# 웹 자동화 (향후 스크래핑용)
playwright>=1.48.0

# 유틸리티
python-dotenv>=1.0.0
requests>=2.32.0
python-dateutil>=2.9.0

# 테스트
pytest>=8.3.0
pytest-asyncio>=0.24.0
coverage>=7.6.0

# 대시보드
streamlit>=1.39.0
pyperclip>=1.8.2

# 개발 도구
black>=24.10.0
isort>=5.13.0
flake8>=7.1.0